+++ fava/application.py
@@ -39,13 +39,13 @@ from werkzeug.utils import secure_filename
 
 from fava import LANGUAGES
 from fava import template_filters
+from fava.context import g
 from fava.core import FavaLedger
 from fava.core.charts import FavaJSONEncoder
-from fava.core.helpers import FavaAPIException
 from fava.help import HELP_PAGES
+from fava.helpers import FavaAPIException
 from fava.json_api import json_api
 from fava.serialisation import serialise
-from fava.context import g
 from fava.util import resource_path
 from fava.util import send_file_inline
 from fava.util import setup_logging
