+++ fava/core/__init__.py
@@ -52,14 +52,14 @@ from fava.core.file import get_entry_slice
 from fava.core.filters import AccountFilter
 from fava.core.filters import AdvancedFilter
 from fava.core.filters import TimeFilter
-from fava.core.helpers import BeancountError
-from fava.core.helpers import FavaAPIException
 from fava.core.ingest import IngestModule
 from fava.core.misc import FavaMisc
 from fava.core.number import DecimalFormatModule
 from fava.core.query_shell import QueryShell
 from fava.core.tree import Tree
 from fava.core.watcher import Watcher
+from fava.helpers import BeancountError
+from fava.helpers import FavaAPIException
 from fava.util import date
 from fava.util import pairwise
